new views bugs fixed

+++ app/src/main/java/com/ubung/tc/ubungmobile/controlador/ChooseSportActivity.java
@@ -83,7 +83,7 @@ public class ChooseSportActivity extends FragmentActivity {
 
     }
 
-    public void intentDescription(int id_deporte, long id, String usuario) {
+    public void intentDescription(long id_deporte, long id, String usuario) {
         finish();
         Intent t = new Intent(this, DescriptionSportActivity.class);
         t.putExtra(POSITION, id_deporte + "");
